requirements.txt
bs4
requests
requests-future
flask
pytest
pytest-cov
gunicorn
